removed litedb models
