--- app/Decsys/Data/Entities/LiteDb/UserWordlist.cs
@@ -1,8 +1,0 @@
-using Decsys.Models.Wordlist;
-
-namespace Decsys.Data.Entities.LiteDb;
-
-public class UserWordlist : BaseUserWordlist
-{
-    public List<WordlistRules> Rules { get; set; } = new();
-}
