+++ app/Decsys/Mapping/WordlistMaps.cs
@@ -15,22 +15,14 @@ public class WordlistMaps : Profile
         CreateMap<Data.Entities.Mongo.WordlistRules, WordlistRules>()
              .ForMember(dest => dest.Value, opt => opt.ConvertUsing(new MongoBsonJTokenConverter()));
 
-        CreateMap<WordlistRules, Data.Entities.LiteDb.WordlistRules>()
-             .ForMember(dest => dest.Value, opt => opt.ConvertUsing(new JTokenLiteDbBsonConverter()));
-
-        CreateMap<Data.Entities.LiteDb.WordlistRules, WordlistRules>()
-             .ForMember(dest => dest.Value, opt => opt.ConvertUsing(new LiteDbBsonJTokenConverter()));
-
         CreateMap<WordlistOptions, Data.Entities.WordlistOptions>();
         CreateMap<Data.Entities.WordlistOptions, WordlistOptions>();
-
 
         CreateMap<WordlistWord, Data.Entities.WordlistWord>();
         CreateMap<Data.Entities.WordlistWord, WordlistWord>();
 
-
-        CreateMap<UserWordlist, Data.Entities.UserWordlist>();
-        CreateMap<Data.Entities.UserWordlist, UserWordlist>();
+        CreateMap<UserWordlist, Data.Entities.Mongo.UserWordlist>();
+        CreateMap<Data.Entities.Mongo.UserWordlist, UserWordlist>();
 
     }
 }
